credit_scoring_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
scikit-learn==1.6.0
lightgbm==4.3.0
xgboost==2.0.3
catboost==1.2.3
pandas==2.2.3
numpy==2.2.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-dotenv==1.0.1
pydantic==2.10.0
pydantic-settings==2.6.0
joblib==1.4.2
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.28.0
slowapi==0.1.9
email-validator==2.2.0
shap==0.46.0
streamlit==1.40.0
plotly==5.24.0
passlib[bcrypt]==1.7.4
redis==5.0.1
